Debug Code and Camera
Added a method to draw a Vector3 to the debug screen, helps with camera
stuff. (Camera is 95% right, only weird on verticle up or down
positions).
Also moved the Debug draw code after all other drawing so it shows ontop
of objects instead of behind.
Cleaned up a tiny bit of the camera code (not much)